pyproject.toml
[project]
name = "bluepopcorn"
version = "0.1.0"
description = "iMessage bot for Seerr media requests"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bluepopcorn"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
